backend/requirements.txt
# ============================================================
# NexClip Ecosystem — Python Requirements
# Python 3.10–3.12 | Windows / Linux / macOS
# Auto-validated against venv on 2026-04-14
# ============================================================

# ── Web Framework ────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3
python-multipart==0.0.18

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
greenlet==3.3.2

# ── Task Queue (Celery + Redis) ───────────────────────────────
celery[redis]==5.4.0
redis==5.2.1
kombu==5.6.2
amqp==5.3.1
billiard==4.2.4

# ── Authentication & Security ─────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# --- AI / ML ---
openai==1.58.1
google-genai>=1.0.0
anthropic>=0.40.0

# ── Speech-to-Text (STT) ─────────────────────────────────────
# ElevenLabs Scribe (primary cloud STT)
elevenlabs>=1.0.0
# faster-whisper (local GPU STT fallback)
faster-whisper>=1.1.1
# PyTorch for faster-whisper CUDA inference
torch>=2.0.0
torchaudio>=2.0.0
# CTranslate2 (faster-whisper runtime)
ctranslate2>=4.0.0
# Hugging Face Hub (model downloads)
huggingface_hub>=1.0.0
# WebRTC VAD for silence detection
webrtcvad-wheels>=2.0.14
# Sound device for audio I/O
sounddevice>=0.4.6

# ── Video & Audio Processing ──────────────────────────────────
# ffmpeg Python bindings
ffmpeg-python==0.2.0
# imageio-ffmpeg provides bundled ffmpeg binary
imageio-ffmpeg>=0.5.1
imageio>=2.30.0
# MoviePy video editing
moviepy==1.0.3
proglog>=0.1.10
decorator>=4.4.2
# OpenCV (headless — no GUI)
opencv-python-headless>=4.8.0
# YOLOv8 for speaker/object detection
ultralytics>=8.0.0
# MediaPipe for face/body tracking
mediapipe>=0.10.0
# yt-dlp for video download
yt-dlp>=2026.3.17
# Audio processing
av>=9.0.0
scipy>=1.10.0
# PyAV dependencies
mutagen>=1.46.0

# ── Image & Caption Rendering ─────────────────────────────────
# Pillow for ImageDraw/ImageFont caption rendering
Pillow>=10.0.0
# Font utilities
fonttools>=4.40.0

# ── Scraping (Nexearch) ───────────────────────────────────────
# Apify cloud scraping platform
apify-client>=1.7.0
# Playwright browser automation (for fallback direct scraping)
playwright>=1.40.0

# ── Vector Embeddings & Search (Nexearch) ────────────────────
# Sentence Transformers for DNA embeddings
sentence-transformers>=2.2.0
# ChromaDB for in-memory/disk vector storage
chromadb>=0.5.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2.10.4
pydantic-settings>=2.7.1
httpx==0.28.1
httpcore>=1.0.0
requests>=2.32.3
aiofiles==24.1.0
aiohttp>=3.9.0
anyio>=4.0.0
psutil>=5.9.0
packaging>=24.0
PyYAML>=6.0.0
rich>=13.0.0
tenacity>=8.0.0
colorama>=0.4.6

# ── Data Science Utilities ────────────────────────────────────
numpy>=1.26.0
scipy>=1.10.0
matplotlib>=3.7.0

# ── Logging ──────────────────────────────────────────────────
loguru==0.7.3

# ── Testing (optional, dev-only) ─────────────────────────────
# pytest>=7.0.0
# pytest-asyncio>=0.23.0
# httpx  (already listed above — used for TestClient)
